requirements-engine.txt
numpy
pillow
opencv-python
torch
torchvision
pyinstaller
